moved Updates and Logs under About
